requirements.txt
azure-functions
requests
beautifulsoup4
python-docx
markdownify
python-dotenv
